+++ packages/@expo/fingerprint/src/Fingerprint.types.ts
@@ -67,17 +67,20 @@ export type Platform = 'android' | 'ios';
 
 export interface Options {
   /**
-   * Limit native files to those for specified platforms. Default is `['android', 'ios']`.
+   * Limit native files to those for specified platforms.
+   * @default ['android', 'ios']
    */
   platforms?: Platform[];
 
   /**
-   * I/O concurrency limit. Default is the number of CPU cores.
+   * I/O concurrency limit.
+   * @default The number of CPU cores.
    */
   concurrentIoLimit?: number;
 
   /**
-   * The algorithm to use for `crypto.createHash()`. Default is `'sha1'`.
+   * The algorithm to use for `crypto.createHash()`.
+   * @default 'sha1'
    */
   hashAlgorithm?: string;
 
@@ -119,7 +122,7 @@ export interface Options {
   enableReactImportsPatcher?: boolean;
 
   /**
-   * Use the react-native core autolinking sources from expo-modules-autolinking rather than @react-native-community/cli.
+   * Use the react-native core autolinking sources from `expo-modules-autolinking` rather than `@react-native-community/cli`.
    * @default true for Expo SDK 52 and higher.
    */
   useRNCoreAutolinkingFromExpo?: boolean;
@@ -204,7 +207,7 @@ export interface HashSourceFile {
   filePath: string;
 
   /**
-   * Reasons of this source coming from
+   * Reasons of this source coming from.
    */
   reasons: string[];
 }
@@ -214,7 +217,7 @@ export interface HashSourceDir {
   filePath: string;
 
   /**
-   * Reasons of this source coming from
+   * Reasons of this source coming from.
    */
   reasons: string[];
 }
@@ -225,7 +228,7 @@ export interface HashSourceContents {
   contents: string | Buffer;
 
   /**
-   * Reasons of this source coming from
+   * Reasons of this source coming from.
    */
   reasons: string[];
 }
@@ -235,7 +238,7 @@ export type HashSource = HashSourceFile | HashSourceDir | HashSourceContents;
 export interface DebugInfoFile {
   path: string;
   hash: string;
-  /** Indicates whether the source is transformed by `fileHookTransform` */
+  /** Indicates whether the source is transformed by `fileHookTransform`. */
   isTransformed?: boolean;
 }
 
@@ -247,7 +250,7 @@ export interface DebugInfoDir {
 
 export interface DebugInfoContents {
   hash: string;
-  /** Indicates whether the source is transformed by `fileHookTransform` */
+  /** Indicates whether the source is transformed by `fileHookTransform`. */
   isTransformed?: boolean;
 }