+++ packages/eslint-config-expo/default.js
@@ -1,44 +1,45 @@
-const coreConfig = require('eslint-config-expo/utils/core.js');
-const expoConfig = require('eslint-config-expo/utils/expo.js');
-const reactConfig = require('eslint-config-expo/utils/react.js');
-const typescriptConfig = require('eslint-config-expo/utils/typescript.js');
-const { allExtensions } = require('eslint-config-expo/utils/extensions.js');
+const {
+  jsExtensions,
+  tsExtensions,
+  platformSubextensions,
+  computeExpoExtensions,
+} = require('./utils/extensions');
 
-module.exports = [
-  ...coreConfig,
-  ...typescriptConfig,
-  ...reactConfig,
-  ...expoConfig,
-  {
-    settings: {
-      'import/extensions': allExtensions,
-      'import/resolver': {
-        node: { extensions: allExtensions },
-      },
-    },
-    languageOptions: {
-      globals: {
-        __DEV__: 'readonly',
-        ErrorUtils: false,
-        FormData: false,
-        XMLHttpRequest: false,
-        alert: false,
-        cancelAnimationFrame: false,
-        cancelIdleCallback: false,
-        clearImmediate: false,
-        fetch: false,
-        navigator: false,
-        process: false,
-        requestAnimationFrame: false,
-        requestIdleCallback: false,
-        setImmediate: false,
-        window: false,
-        'shared-node-browser': true,
-      },
+const allExtensions = computeExpoExtensions(
+  [...jsExtensions, ...tsExtensions],
+  platformSubextensions
+);
+
+module.exports = {
+  extends: ['./utils/core.js', './utils/typescript.js', './utils/react.js', './utils/expo.js'],
+  globals: {
+    __DEV__: 'readonly',
+    ErrorUtils: false,
+    FormData: false,
+    XMLHttpRequest: false,
+    alert: false,
+    cancelAnimationFrame: false,
+    cancelIdleCallback: false,
+    clearImmediate: false,
+    fetch: false,
+    navigator: false,
+    process: false,
+    requestAnimationFrame: false,
+    requestIdleCallback: false,
+    setImmediate: false,
+    window: false,
+    'shared-node-browser': true,
+  },
+  settings: {
+    'import/extensions': allExtensions,
+    'import/resolver': {
+      node: { extensions: allExtensions },
     },
   },
-  {
-    files: ['*.web.*'],
-    env: { browser: true },
-  },
-];
+  overrides: [
+    {
+      files: ['*.web.*'],
+      env: { browser: true },
+    },
+  ],
+};