+++ packages/eslint-config-expo/utils/expo.js
@@ -1,19 +1,12 @@
-const expo = require('eslint-plugin-expo');
-
-module.exports = [
-  {
+module.exports = {
+  plugins: ['expo'],
+  ignorePatterns: [
     // JS files can end up in build intermediates, eg:
     // android/app/build/intermediates/assets/debug/EXDevMenuApp.android.js
-    ignores: ['android/app/build'],
+    'android/app/build',
+  ],
+  rules: {
+    'expo/no-env-var-destructuring': ['error'],
+    'expo/no-dynamic-env-var': ['error'],
   },
-  {
-    plugins: {
-      expo,
-    },
-
-    rules: {
-      'expo/no-env-var-destructuring': ['error'],
-      'expo/no-dynamic-env-var': ['error'],
-    },
-  },
-];
+};