+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -73,6 +73,7 @@ import expo.modules.interfaces.barcodescanner.BarCodeScannerResult.BoundingBox
 import expo.modules.interfaces.camera.CameraViewInterface
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.Promise
+import expo.modules.kotlin.RuntimeContext
 import expo.modules.kotlin.exception.Exceptions
 import expo.modules.kotlin.viewevent.EventDispatcher
 import expo.modules.kotlin.views.ExpoView
@@ -83,6 +84,7 @@ import kotlinx.coroutines.launch
 import java.io.File
 import kotlin.math.roundToInt
 import kotlin.properties.Delegates
+import kotlin.text.toFloat
 
 const val ANIMATION_FAST_MILLIS = 50L
 const val ANIMATION_SLOW_MILLIS = 100L
@@ -201,6 +203,9 @@ class ExpoCameraView(
     setTorchEnabled(newValue)
   }
 
+  private var lastWidth = 0
+  private var lastHeight = 0
+
   private val onCameraReady by EventDispatcher<Unit>()
   private val onMountError by EventDispatcher<CameraMountErrorEvent>()
   private val onBarcodeScanned by EventDispatcher<BarcodeScannedEvent>(
@@ -236,8 +241,12 @@ class ExpoCameraView(
   override fun onLayout(changed: Boolean, left: Int, top: Int, right: Int, bottom: Int) {
     val width = right - left
     val height = bottom - top
-    previewView.layout(0, 0, width, height)
-    glSurfaceTexture?.setDefaultBufferSize(width, height)
+    if (width != lastWidth || height != lastHeight) {
+      previewView.layout(0, 0, width, height)
+      glSurfaceTexture?.setDefaultBufferSize(width, height)
+      lastWidth = width
+      lastHeight = height
+    }
   }
 
   override fun onViewAdded(child: View?) {
@@ -250,7 +259,7 @@ class ExpoCameraView(
     addView(previewView, 0)
   }
 
-  fun takePicture(options: PictureOptions, promise: Promise, cacheDirectory: File) {
+  fun takePicture(options: PictureOptions, promise: Promise, cacheDirectory: File, runtimeContext: RuntimeContext) {
     val audioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager
     val volume = audioManager.getStreamVolume(AudioManager.STREAM_MUSIC)
     val hasShutterSound = options.shutterSound
@@ -280,11 +289,14 @@ class ExpoCameraView(
           if (options.fastMode) {
             promise.resolve(null)
           }
+
           cacheDirectory.let {
             scope.launch {
               val shouldMirror = mirror && lensFacing == CameraType.FRONT
-              ResolveTakenPicture(data, promise, options, shouldMirror, it) { response: Bundle ->
-                onPictureSaved(response)
+              ResolveTakenPicture(data, promise, options, shouldMirror, runtimeContext, it) { response: Bundle ->
+                if (!options.pictureRef) {
+                  onPictureSaved(response)
+                }
               }.resolve()
             }
           }
@@ -398,12 +410,14 @@ class ExpoCameraView(
       {
         val cameraProvider: ProcessCameraProvider = providerFuture.get()
 
-        previewView.scaleType =
-          if (ratio == CameraRatio.FOUR_THREE || ratio == CameraRatio.SIXTEEN_NINE) {
-            PreviewView.ScaleType.FIT_CENTER
-          } else {
-            PreviewView.ScaleType.FILL_CENTER
-          }
+        ratio?.let {
+          previewView.scaleType =
+            if (ratio == CameraRatio.FOUR_THREE || ratio == CameraRatio.SIXTEEN_NINE) {
+              PreviewView.ScaleType.FIT_CENTER
+            } else {
+              PreviewView.ScaleType.FILL_CENTER
+            }
+        }
 
         val resolutionSelector = buildResolutionSelector()
         val preview = Preview.Builder()
@@ -457,7 +471,7 @@ class ExpoCameraView(
           // Set the previous zoom level after recreating the camera
           camera?.cameraControl?.setLinearZoom(zoom.coerceIn(0f, 1f))
           this.cameraProvider = cameraProvider
-        } catch (e: Exception) {
+        } catch (_: Exception) {
           onMountError(
             CameraMountErrorEvent("Camera component could not be rendered - is there any other instance running?")
           )
@@ -616,11 +630,6 @@ class ExpoCameraView(
     (context.getSystemService(Context.WINDOW_SERVICE) as WindowManager).defaultDisplay.rotation
   }
 
-  fun releaseCamera() = appContext.mainQueue.launch {
-    shouldCreateCamera = true
-    cameraProvider?.unbindAll()
-  }
-
   private fun transformBarcodeScannerResultToViewCoordinates(barcode: BarCodeScannerResult) {
     val cornerPoints = barcode.cornerPoints
     val previewWidth = previewView.width
@@ -739,9 +748,17 @@ class ExpoCameraView(
     onPictureSaved(PictureSavedEvent(response.getInt("id"), response.getBundle("data")!!))
   }
 
-  fun cancelCoroutineScope() = try {
+  private fun cancelCoroutineScope() = try {
     scope.cancel(ModuleDestroyedException())
   } catch (e: Exception) {
     Log.e(CameraViewModule.TAG, "The scope does not have a job in it")
+  }
+
+  override fun onDetachedFromWindow() {
+    super.onDetachedFromWindow()
+    orientationEventListener.disable()
+    cancelCoroutineScope()
+    cameraProvider?.unbindAll()
+    glSurfaceTexture?.release()
   }
 }