+++ packages/expo-camera/ios/Current/CameraView.swift
@@ -134,7 +134,9 @@ public class CameraView: ExpoView, EXAppLifecycleListener,
   let onResponsiveOrientationChanged = EventDispatcher()
 
   private var deviceOrientation: UIInterfaceOrientation {
-    window?.windowScene?.interfaceOrientation ?? .unknown
+    UIApplication.shared.connectedScenes.compactMap {
+      $0 as? UIWindowScene
+    }.first?.interfaceOrientation ?? .unknown
   }
 
   required init(appContext: AppContext? = nil) {
@@ -418,9 +420,8 @@ public class CameraView: ExpoView, EXAppLifecycleListener,
       return
     }
 
-    let imageData = photo.fileDataRepresentation()
     handleCapturedImageData(
-      imageData: imageData,
+      imageData: photo.fileDataRepresentation(),
       metadata: photo.metadata,
       options: options,
       promise: promise
@@ -455,12 +456,6 @@ public class CameraView: ExpoView, EXAppLifecycleListener,
     let croppedSize = AVMakeRect(aspectRatio: previewSize, insideRect: cropRect)
 
     takenImage = ExpoCameraUtils.crop(image: takenImage, to: croppedSize)
-
-    let path = FileSystemUtilities.generatePathInCache(
-      appContext,
-      in: "Camera",
-      extension: ".jpg"
-    )
 
     let width = takenImage.size.width
     let height = takenImage.size.height
@@ -522,9 +517,22 @@ public class CameraView: ExpoView, EXAppLifecycleListener,
     }
 
     guard let processedImageData else {
-      promise.reject(CameraSavingImageException())
+      promise.reject(CameraSavingImageException("Image data could not be processed"))
       return
     }
+
+    if options.pictureRef {
+      if let image = UIImage(data: processedImageData) {
+        promise.resolve(PictureRef(image))
+        return
+      }
+    }
+
+    let path = FileSystemUtilities.generatePathInCache(
+      appContext,
+      in: "Camera",
+      extension: ".jpg"
+    )
 
     response["uri"] = ExpoCameraUtils.write(data: processedImageData, to: path)
     response["width"] = width